refactor: renames builder folder to user
