+++ src/app/collective-rewards/user/components/index.ts
