+++ src/app/collective-rewards/user/index.ts
